refactor(document-preview): extract facade metadata rendering and shared helpers
